delete user controller

+++ src/controllers/user/DeactivateUser/index.ts
@@ -1,24 +1,32 @@
 import { Application, Request, Response } from 'express'
+import mongoose from 'mongoose'
 
-import { ApiUrlPath } from 'shared/utils/index'
+import { ApiUrlPath, AppError } from 'shared/utils/index'
 import { userAuthMiddleware } from 'middleware/index'
 import { UserModel, QuestionModel, AnswerModel } from 'models/index'
 
+const ObjectId = mongoose.Types.ObjectId
+
 export default (app: Application) =>
-  app.get(ApiUrlPath.DeactivateUser, userAuthMiddleware, (req: Request, res: Response) => {
+  app.delete(ApiUrlPath.DeactivateUser, userAuthMiddleware, (req: Request, res: Response) => {
     //
     const { APP_URL } = process.env
-    const userId = req.decoded._id
+    const userId = ObjectId(req.decoded._id)
 
-    AnswerModel.deleteMany({ answererId: userId })
-      .then(() => {
-        QuestionModel.deleteMany({ creatorId: userId })
-          .then(() => {
-            UserModel.deleteOne({ _id: req.decoded._id })
-              .then(() => res.redirect(APP_URL + '?logout'))
-              .catch(err => res.status(400).send(err.errors))
-          })
-          .catch(err => res.status(400).send(err.errors))
+    UserModel.deleteOne({ _id: userId })
+      .then(({ deletedCount }) => {
+        if (deletedCount === 1) {
+          Promise.all([
+            QuestionModel.deleteMany({ creatorId: userId }),
+            AnswerModel.deleteMany({ answererId: userId }),
+          ])
+            .then(() => {
+              res.redirect(301, APP_URL + '?logout')
+            })
+            .catch(err => res.status(400).send(err.errors))
+        } else {
+          res.status(400).send(AppError.NoSuchUserError)
+        }
       })
       .catch(err => res.status(400).send(err.errors))
   })
